# templates/service/internal/model/model.go
// TEMPLATE: Internal domain models for __SERVICE_FULL__
// Copy this file and replace all __PLACEHOLDER__ markers with your service values.
//
// Internal models represent the canonical shape of data within the service.
// They are NOT the same as the public API types (pkg/types) — the separation
// allows the internal representation to evolve independently of the wire format.
//
// Conventions:
//   - Internal models live in internal/model and are never imported outside
//     this module.
//   - Public API types live in pkg/types and are imported by CLI, client SDK,
//     and other consumers.
//   - Conversion between internal and public types happens in the server
//     handlers (toPublic__RESOURCE__, fromPublicCreate__RESOURCE__, etc.).
//   - Internal models carry database-level fields (e.g., primary key, audit
//     timestamps) that may not appear in the public types.
package model

import "time"

// __RESOURCE__ is the internal domain representation of a __RESOURCE_LOWER__ resource.
//
// TEMPLATE: Add, rename, or remove fields to match your domain model.
// Keep this struct aligned with the database schema columns.
type __RESOURCE__ struct {
	// ID is the primary key (UUID). Generated by the database if not provided.
	ID string

	// Name is a human-readable label for the __RESOURCE_LOWER__.
	Name string

	// Description is an optional longer description.
	Description string

	// TEMPLATE: Add domain-specific fields here. Examples:
	//
	// Type is the __RESOURCE_LOWER__ classification.
	// Type string
	//
	// Status is the operational state (e.g., "active", "inactive").
	// Status string
	//
	// ParentID references an optional parent resource.
	// ParentID *string
	//
	// Tags is a set of key-value labels attached to the resource.
	// Tags map[string]string
	//
	// Arch is the hardware architecture (e.g., "x86_64", "aarch64").
	// Arch string

	// CreatedAt is the row insertion timestamp (set by the database).
	CreatedAt time.Time

	// UpdatedAt is the last modification timestamp (set by the database).
	UpdatedAt time.Time
}

// TEMPLATE: Add additional domain model types below.
//
// Example: If the service manages a secondary resource linked to __RESOURCE__:
//
// type __RESOURCE__Attribute struct {
//     ID           string
//     __RESOURCE__ID string
//     Key          string
//     Value        string
//     CreatedAt    time.Time
//     UpdatedAt    time.Time
// }